requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Processing
numpy==1.26.4
pandas==2.2.1

# Embeddings and Vector Store
openai==1.12.0
sentence-transformers==2.5.1
faiss-cpu==1.8.0

# LLM and RAG
langchain==0.1.10
langchain-openai==0.0.8
langchain-community==0.0.24

# API
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9

# Web UI
streamlit==1.31.1

# Configuration
python-dotenv==1.0.1

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
